Refactor: move mail block to watch page
